refactor: split atmosphere editor group to own module
body_appearance.py had grown to 1057 lines, over the 1000-line
module cap. Move the atmosphere (haze) group builder and its three
handlers into a sibling body_atmosphere.py mixin, following the
ring_advanced/ring_tab pattern: SimEditorBase declares the _body /
_body_tab_widget / _build_atmosphere_group seams, and the main
window composes the new mixin. Both modules are now under the cap
(917 and 169 lines); widget references stay on the tab widget, so
the round-trip and per-widget tests reach every key unchanged.
